# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NavorTech fleet telemetry read API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type DeviceModel = (typeof DeviceModel)[keyof typeof DeviceModel];

export const DeviceModel = {
  VL06: "VL06",
  VL08: "VL08",
  unknown: "unknown",
} as const;

export type Transport = (typeof Transport)[keyof typeof Transport];

export const Transport = {
  udp: "udp",
  tcp: "tcp",
} as const;

export type ParseStatus = (typeof ParseStatus)[keyof typeof ParseStatus];

export const ParseStatus = {
  ok: "ok",
  no_frames: "no_frames",
  checksum_mismatch: "checksum_mismatch",
  no_trailer: "no_trailer",
  response_to_command: "response_to_command",
  duplicate: "duplicate",
} as const;

export type Direction = (typeof Direction)[keyof typeof Direction];

export const Direction = {
  device_to_server: "device_to_server",
  server_to_device: "server_to_device",
} as const;

/**
 * Tracker registry row with denormalized latest-state snapshot.
All `last*` fields are nullable until the first successful report.

 */
export interface Device {
  id: string;
  model: DeviceModel;
  firstSeenAt: string;
  lastSeenAt: string;
  lastPeer?: string | null;
  lastTransport?: Transport | null;
  lastLat?: number | null;
  lastLon?: number | null;
  lastIgnition?: boolean | null;
  lastSpeedKmh?: number | null;
  lastHourmeterMin?: number | null;
  lastOdometerM?: number | null;
  lastDeviceTs?: string | null;
  lastReportAt?: string | null;
  /** Human-readable asset name (e.g. "BOB-001") */
  displayName?: string | null;
  /** Equipment model name (e.g. "Bobcat E35") */
  machineModel?: string | null;
  /** Equipment type (e.g. "Mini Escavadeira") */
  machineType?: string | null;
  /** Service interval in hours (e.g. 500) */
  serviceLimitHours?: number | null;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

/**
 * Fields that can be patched on a device by an operator
 */
export interface DeviceUpdate {
  displayName?: string | null;
  machineModel?: string | null;
  machineType?: string | null;
  serviceLimitHours?: number | null;
  notes?: string | null;
}

/**
 * Raw inbound packet (UDP datagram or TCP framed boundary).
 */
export interface Packet {
  id: string;
  receivedAt: string;
  peer: string;
  transport: Transport;
  bytes: number;
  ascii: string;
  parseStatus: ParseStatus;
  frameCount: number;
  deviceId?: string | null;
  msgnum?: string | null;
  ackedAt?: string | null;
  ackAscii?: string | null;
}

export interface ReportCommon {
  id: string;
  frameId: string;
  deviceId: string;
  eventIndex?: number | null;
  protocolId?: string | null;
  deviceTs?: string | null;
  receivedAt: string;
  rawBody: string;
}

export type ReportRgp = ReportCommon & {
  lat?: number | null;
  lon?: number | null;
  speedKmh?: number | null;
  headingDeg?: number | null;
  gpsStatus?: number | null;
  secondsSinceFix?: number | null;
  hdop?: number | null;
  diRaw?: number | null;
  ignition?: boolean | null;
  mainPower?: boolean | null;
};

export type ReportRuv00 = ReportCommon & {
  backupBatteryV?: number | null;
  mainSupplyV?: number | null;
  serial?: string | null;
  firmware?: string | null;
  board?: string | null;
  script?: string | null;
  peripheral?: string | null;
  rainSpeedFlag?: number | null;
  hourmeterSource?: number | null;
  sharpEventSource?: number | null;
  iccid?: string | null;
};

export type ReportRuv01 = ReportCommon & {
  lat?: number | null;
  lon?: number | null;
  speedKmh?: number | null;
  headingDeg?: number | null;
  gpsStatus?: number | null;
  secondsSinceFix?: number | null;
  hdop?: number | null;
  diRaw?: number | null;
  ignition?: boolean | null;
  mainPower?: boolean | null;
  backupBatteryV?: number | null;
  mainSupplyV?: number | null;
  maxSpeedViolation?: number | null;
  maxRpmViolation?: number | null;
  hourmeterMin?: number | null;
  odometerM?: number | null;
  rpm?: number | null;
  engineTempC?: number | null;
  oilPressureKpa?: number | null;
  fuelPct?: number | null;
  rainSpeedFlag?: number | null;
  online?: boolean | null;
  network?: string | null;
  driverId?: string | null;
};

export type ReportRuv02 = ReportCommon & {
  rpmRange1Sec?: number | null;
  rpmRange2Sec?: number | null;
  rpmRange3Sec?: number | null;
  rpmRange4Sec?: number | null;
  rpmRange5Sec?: number | null;
  inertialSec?: number | null;
  engineBrakingSec?: number | null;
  coastingSec?: number | null;
  travelTimeMin?: number | null;
  distanceM?: number | null;
  fuelConsumedDecilitres?: number | null;
};

export type ReportRuv03 = ReportCommon & {
  acceleratorPct?: number | null;
  hourmeterMin?: number | null;
  odometerM?: number | null;
  rpm?: number | null;
  engineTempC?: number | null;
  enginePressureKpa?: number | null;
  fuelPct?: number | null;
  cumulativeFuelDecilitres?: number | null;
  speedKmh?: number | null;
  engineTorquePct?: number | null;
  engineBrakePct?: number | null;
  cruiseControl?: boolean | null;
  clutch?: boolean | null;
  parkingBrake?: boolean | null;
  serviceBrake?: boolean | null;
};

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;

export type DeviceModelFilterParameter = DeviceModel;

/**
 * Page size (1-200, default 50)
 */
export type LimitParameter = number;

/**
 * ISO timestamp cursor — return rows with `received_at < before`
 */
export type BeforeParameter = string;

export type ParseStatusFilterParameter = ParseStatus;

export type ListDevicesParams = {
  /**
   * Filter by device model
   */
  model?: DeviceModelFilterParameter;
};

export type ListDevicePacketsParams = {
  /**
   * Page size (1-200, default 50)
   * @minimum 1
   * @maximum 200
   */
  limit?: LimitParameter;
  /**
   * ISO timestamp cursor — return rows with `received_at < before`
   */
  before?: BeforeParameter;
  /**
   * Filter packets by parse outcome
   */
  parseStatus?: ParseStatusFilterParameter;
};

export type ListDeviceReportsRgpParams = {
  /**
   * Page size (1-200, default 50)
   * @minimum 1
   * @maximum 200
   */
  limit?: LimitParameter;
  /**
   * ISO timestamp cursor — return rows with `received_at < before`
   */
  before?: BeforeParameter;
};

export type ListDeviceReportsRuv00Params = {
  /**
   * Page size (1-200, default 50)
   * @minimum 1
   * @maximum 200
   */
  limit?: LimitParameter;
  /**
   * ISO timestamp cursor — return rows with `received_at < before`
   */
  before?: BeforeParameter;
};

export type ListDeviceReportsRuv01Params = {
  /**
   * Page size (1-200, default 50)
   * @minimum 1
   * @maximum 200
   */
  limit?: LimitParameter;
  /**
   * ISO timestamp cursor — return rows with `received_at < before`
   */
  before?: BeforeParameter;
};

export type ListDeviceReportsRuv02Params = {
  /**
   * Page size (1-200, default 50)
   * @minimum 1
   * @maximum 200
   */
  limit?: LimitParameter;
  /**
   * ISO timestamp cursor — return rows with `received_at < before`
   */
  before?: BeforeParameter;
};

export type ListDeviceReportsRuv03Params = {
  /**
   * Page size (1-200, default 50)
   * @minimum 1
   * @maximum 200
   */
  limit?: LimitParameter;
  /**
   * ISO timestamp cursor — return rows with `received_at < before`
   */
  before?: BeforeParameter;
};

export type ListFleetRecentPacketsParams = {
  /**
   * Page size (1-200, default 50)
   * @minimum 1
   * @maximum 200
   */
  limit?: LimitParameter;
  /**
   * ISO timestamp cursor — return rows with `received_at < before`
   */
  before?: BeforeParameter;
  /**
   * Filter packets by parse outcome
   */
  parseStatus?: ParseStatusFilterParameter;
};